# Conflicts:
#   backend/src/main/java/site/sleepmate/backend/repository/VideoOrderRepository.java
#   backend/src/main/java/site/sleepmate/backend/repository/VideoRecordRepository.java

+++ backend/src/main/java/site/sleepmate/backend/repository/VideoOrderRepository.java
@@ -3,5 +3,6 @@ package site.sleepmate.backend.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import site.sleepmate.backend.domain.VideoOrder;
 
-public interface VideoOrderRepository extends JpaRepository<VideoOrder, Integer> {
+public interface VideoOrderRepository extends JpaRepository<VideoOrder, Long> {
+
 }
